requirements.txt
# Astro Runtime includes the following pre-installed providers packages: 
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos
apache-airflow-providers-snowflake

# MySQL Support for Apache Airflow
apache-airflow-providers-mysql  # MySQL provider for Airflow

# Additional dependencies
requests  # For API calls
snowflake-connector-python  # Snowflake Python connector
mysql-connector-python  # MySQL Python connector

apache-airflow-providers-apprise==2.0.0
